Two new keyboard shortcuts and a whole refactoring
