+++ src/message.rs
@@ -1,0 +1,37 @@
+use crate::xmpp;
+
+#[derive(Debug, Clone)]
+pub enum Message
+{
+    Ignore,
+    TabPressed,
+    ShiftTabPressed,
+    NextSelection,
+    PrevSelection,
+    JidInputChanged(String),
+    PasswordInputChanged(String),
+    RememberMeToggled(bool),
+    SaveRoomToggled(bool),
+    FocusPassword,
+    Connect,
+    CancelConnect,
+    XmppEvent(xmpp::XmppEvent),
+    Disconnect,
+    SelectRoom(usize),
+    SelectChat(usize),
+    StartChat(String),
+    InputChanged(String),
+    SendMessage,
+    ShowJoinPanel,
+    HideJoinPanel,
+    JoinInputChanged(String),
+    JoinRoom,
+    DismissJoinError,
+    LeaveRoom,
+    CloseChat,
+    LeaveSelection,
+    OpenUrl(String),
+    ForgetAutoLogin,
+    WindowFocused,
+    WindowUnfocused,
+}
